Record the Rust tracking issue for `target_has_atomic` inline
Note that this patch is a minimal-effort basis to enable `radium` to
target specifically `thumbv7m-none-eabi`, and is not generalized to
support all Rust targets. It may need revisited if other issues are
filed for absent symbols, but I consider it better to do the minimal
work needed today than to produce an elaborate, brittle, work-around
that is not actively required and will, ideally, be rendered
obsolete in a future release of the compiler.