Add type aliases configured by the build script
Since `radium` cannot export macros capable of precise target
detection, and build-script `cfg` entries are visible only to the
crate to which the build script belongs, `radium` target detection
must be exported some other way. The easiest way to do this is with
a type alias that is conditionally set to the atom when it exists,
and the cell when the atom is unavailable.