Add config for avr arch.
Also, discover the arch in a better way. On the avr arch the target
triple only has two parts. This confuses the detection script in that
it doesn't know how to parse the target string. Instead, I use the
variable that cargo passes that tells the actual arch without parsing.

See https://doc.rust-lang.org/cargo/reference/environment-variables.html
for more info on the variable.